+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Path config is in `mjs` instead of JSON. JSON config is still supported, but migration to ESM encouraged.
 - Support for `mjs` collection files for injection into HTML or generators input. Support for JSON is kept, but ESM is favored.
 - Added `posthtml` step for additional tasks
+- Added `import-wp` task for import of WordPress Posts and Pages
 
 ## 8.7.1